backend/pyproject.toml
[project]
name = "fundi-backend"
version = "0.1.0"
description = "Fundi CRM — backend API"
authors = [{ name = "Samuel Agyei", email = "agyeisterogh@gmail.com" }]
requires-python = ">=3.13,<4.0"
package-mode = false
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-unfold (>=0.89.0,<0.90.0)",
    "celery[redis] (>=5.6.3,<6.0.0)",
    "django-celery-beat (>=2.9.0,<3.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker (>=40.13.0,<41.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "ipython (>=9.12.0,<10.0.0)",
    "ruff (>=0.15.10,<0.16.0)",
    "django-stubs[compatible-mypy] (>=6.0.2,<6.1.0)",
    "djangorestframework-stubs[compatible-mypy] (>=3.16.8,<3.17.0)",
]

[tool.mypy]
python_version = "3.13"
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
files = [
    "apps",
    "config",
]
exclude = [
    "^apps/.*/admin\\.py$",
    "^apps/.*/migrations/.*",
    "^apps/.*/tests/.*",
    "^apps/.*/tests\\.py$",
]
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

[[tool.mypy.overrides]]
module = [
    "decouple",
    "unfold.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 88
extend-exclude = [
    ".venv",
    "staticfiles",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
]

[tool.ruff.lint.per-file-ignores]
"apps/*/migrations/*.py" = ["E501", "I001"]
"config/settings/*.py" = ["F403", "F405"]
